# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SELETIVO",
        "version": "0.1"
    },
    "paths": {
...
        "/api/endereco": {
            "get": {
                "tags": [
                    "Endereço"
                ],
                "summary": "GET Endereços",
                "description": "GET Paginado",
                "operationId": "9c10249b6d54a91b77e519f61d558174",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store-endereco": {
            "post": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Cria um novo endereço",
                "description": "Endpoint para cadasatrar um novo endereço no sistema.",
                "operationId": "4933802f7da27814f755064f72e51aea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "end_tipo_logradouro",
                                    "end_logradouro",
                                    "end_numero",
                                    "end_bairro",
                                    "cid_id"
                                ],
                                "properties": {
                                    "end_tipo_logradouro": {
                                        "type": "string",
                                        "example": "rua ficticia"
                                    },
                                    "end_logradouro": {
                                        "type": "string",
                                        "example": "logradouro"
                                    },
                                    "end_numero": {
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "end_bairro": {
                                        "type": "string",
                                        "example": "CPA II"
                                    },
                                    "cid_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço Cadastrado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço Cadastrado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/show-endereco/{end_id}": {
            "get": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Show Endereço",
                "description": "Show Endereço",
                "operationId": "8b0464b7b5dd7a9ffd66480c2fff0e9d",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "ID do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-endereco/{end_id}": {
            "put": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Atualizar uma Endereco",
                "description": "Endpoint para atualizar um novo endereço",
                "operationId": "378f7ff1666336805c77d3d1cd2048de",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "ID do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "end_tipo_logradouro": {
                                        "type": "string",
                                        "example": "Rua"
                                    },
                                    "end_logradouro": {
                                        "type": "string",
                                        "example": "Rua 15"
                                    },
                                    "end_numero": {
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "end_bairro": {
                                        "type": "string",
                                        "example": "CPA III"
                                    },
                                    "cid_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço atualizado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "end_tipo_logradouro": {
                                                    "type": "string",
                                                    "example": "rua ficticia"
                                                },
                                                "end_logradouro": {
                                                    "type": "string",
                                                    "example": "logradouro"
                                                },
                                                "end_numero": {
                                                    "type": "string",
                                                    "example": "01"
                                                },
                                                "end_bairro": {
                                                    "type": "string",
                                                    "example": "CPA III"
                                                },
                                                "cid_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-endereco/{end_id}": {
            "delete": {
                "tags": [
                    "Endereço"
                ],
                "summary": "Deleta Endereço",
                "description": "Remove Endereço",
                "operationId": "3a3a0c8c1afaf040bc2c329ae1e48638",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "ID do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço Removido",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}